+++ jgiven-core/src/main/java/com/tngtech/jgiven/annotation/As.java
@@ -4,7 +4,7 @@ import java.lang.annotation.*;
 
 /**
  * This annotation can be used to override the default
- * representation for a step method or test method in the report.
+ * representation for a step method, test method or class name in the report.
  * 
  * <p>
  * Note that the '$' character keeps its special meaning and will be
@@ -25,7 +25,8 @@ import java.lang.annotation.*;
  * }
  * </pre>
  * 
- * @since 0.7.4
+ * 
+ * @since 0.7.4. Since 0.9.0 this annotation can be applied to test classes as well
  */
 @Retention( RetentionPolicy.RUNTIME )
 @Target( { ElementType.METHOD, ElementType.TYPE } )