+++ jgiven-core/src/main/java/com/tngtech/jgiven/config/DefaultConfiguration.java
@@ -1,6 +1,6 @@
 package com.tngtech.jgiven.config;
 
-public final class DefaultConfiguration extends AbstractJGivenConfiguraton {
+public final class DefaultConfiguration extends AbstractJGivenConfiguration {
 
     @Override
     public void configure() {}