+++ jgiven-core/src/main/java/com/tngtech/jgiven/format/ArgumentFormatter.java
@@ -1,9 +1,15 @@
 package com.tngtech.jgiven.format;
 
 /**
- * Interface for defining customer argument formatter.
+ * Interface for defining custom argument formatter using the {@link com.tngtech.jgiven.annotation.Format} annotation.
+ * <p>
+ * Note that in many cases it is more convenient to define a custom formatter annotation using the {@link com.tngtech.jgiven.annotation.AnnotationFormat}
+ * annotation and defining an {@link com.tngtech.jgiven.format.AnnotationArgumentFormatter}.
+ * <p>
+ * Alternatively you can also provide a global formatter for a type (see {@link com.tngtech.jgiven.format.Formatter}).
  *
  * @param <T> the type of the object to format
+ * @see com.tngtech.jgiven.annotation.Format
  */
 public interface ArgumentFormatter<T> {
     /**