+++ jgiven-core/src/main/java/com/tngtech/jgiven/impl/ScenarioModelBuilder.java
@@ -26,6 +26,7 @@ import com.tngtech.jgiven.format.Formatter;
 import com.tngtech.jgiven.format.TableFormatter;
 import com.tngtech.jgiven.impl.intercept.ScenarioListener;
 import com.tngtech.jgiven.impl.util.AssertionUtil;
+import com.tngtech.jgiven.impl.util.ReflectionUtil;
 import com.tngtech.jgiven.impl.util.WordUtil;
 import com.tngtech.jgiven.report.model.*;
 import com.tngtech.jgiven.report.model.StepFormatter.Formatting;
@@ -33,11 +34,11 @@ import com.tngtech.jgiven.report.model.StepFormatter.Formatting;
 public class ScenarioModelBuilder implements ScenarioListener {
     private static final Logger log = LoggerFactory.getLogger( ScenarioModelBuilder.class );
     private static final Formatting<?, ?> DEFAULT_FORMATTING = new StepFormatter.ArgumentFormatting<ArgumentFormatter<Object>, Object>(
-            new DefaultFormatter<Object>() );
+        new DefaultFormatter<Object>() );
 
     private static final Set<String> STACK_TRACE_FILTER = ImmutableSet
-            .of( "sun.reflect", "com.tngtech.jgiven.impl.intercept", "com.tngtech.jgiven.impl.intercept", "$$EnhancerByCGLIB$$",
-                    "java.lang.reflect", "net.sf.cglib.proxy", "com.sun.proxy" );
+        .of( "sun.reflect", "com.tngtech.jgiven.impl.intercept", "com.tngtech.jgiven.impl.intercept", "$$EnhancerByCGLIB$$",
+            "java.lang.reflect", "net.sf.cglib.proxy", "com.sun.proxy" );
     private static final boolean FILTER_STACK_TRACE = Config.config().filterStackTrace();
 
     private ScenarioModel scenarioModel;
@@ -150,7 +151,7 @@ public class ScenarioModelBuilder implements ScenarioListener {
     private <T> Formatting<?, ?> getFormatting( Class<T> parameterType, Annotation[] annotations ) {
         Formatting<?, ?> formatting = getFormatting( annotations, Sets.<Class<?>>newHashSet(), null );
         if( formatting == null ) {
-            Formatter<T> formatter = configuration.getFormatter( parameterType );
+            Formatter<T> formatter = (Formatter<T>) configuration.getFormatter( parameterType );
             if( formatter != null ) {
                 formatting = new StepFormatter.TypeBasedFormatting<T>( formatter, annotations );
             } else {
@@ -177,7 +178,7 @@ public class ScenarioModelBuilder implements ScenarioListener {
                 } else if( annotation instanceof AnnotationFormat ) {
                     AnnotationFormat arg = (AnnotationFormat) annotation;
                     return new StepFormatter.ArgumentFormatting(
-                            new StepFormatter.AnnotationBasedFormatter( arg.value().newInstance(), originalAnnotation ) );
+                        new StepFormatter.AnnotationBasedFormatter( arg.value().newInstance(), originalAnnotation ) );
                 } else {
                     Class<? extends Annotation> annotationType = annotation.annotationType();
                     if( !visitedTypes.contains( annotationType ) ) {
@@ -306,6 +307,23 @@ public class ScenarioModelBuilder implements ScenarioListener {
 
         List<Formatting<?, ?>> formatter = getFormatter( method.getParameterTypes(), method.getParameterAnnotations() );
         setArguments( toStringList( formatter, getValues( namedArguments ) ) );
+        setCaseDescription( method, namedArguments );
+    }
+
+    private void setCaseDescription( Method method, List<NamedArgument> namedArguments ) {
+        if( method.isAnnotationPresent( CaseDescription.class ) ) {
+            CaseDescription annotation = method.getAnnotation( CaseDescription.class );
+            CaseDescriptionProvider caseDescriptionProvider = ReflectionUtil.newInstance( annotation.provider() );
+            String value = annotation.value();
+            List<?> values;
+            if( annotation.formatValues() ) {
+                values = scenarioCaseModel.getExplicitArguments();
+            } else {
+                values = getValues( namedArguments );
+            }
+            String caseDescription = caseDescriptionProvider.description( value, scenarioModel.getExplicitParameters(), values );
+            scenarioCaseModel.setDescription( caseDescription );
+        }
     }
 
     private List<Object> getValues( List<NamedArgument> namedArguments ) {
@@ -457,9 +475,9 @@ public class ScenarioModelBuilder implements ScenarioListener {
         String name = Strings.isNullOrEmpty( isTag.name() ) ? isTag.type() : isTag.name();
 
         return TagConfiguration.builder( annotation.annotationType() ).defaultValue( isTag.value() ).description( isTag.description() )
-                .explodeArray( isTag.explodeArray() ).ignoreValue( isTag.ignoreValue() ).prependType( isTag.prependType() ).name( name )
-                .descriptionGenerator( isTag.descriptionGenerator() ).cssClass( isTag.cssClass() ).color( isTag.color() )
-                .style( isTag.style() ).tags( getTagNames( isTag, annotation ) ).build();
+            .explodeArray( isTag.explodeArray() ).ignoreValue( isTag.ignoreValue() ).prependType( isTag.prependType() ).name( name )
+            .descriptionGenerator( isTag.descriptionGenerator() ).cssClass( isTag.cssClass() ).color( isTag.color() )
+            .style( isTag.style() ).tags( getTagNames( isTag, annotation ) ).build();
 
     }
 
@@ -502,8 +520,8 @@ public class ScenarioModelBuilder implements ScenarioListener {
             return tagConfiguration.getDescriptionGenerator().newInstance().generateDescription( tagConfiguration, annotation, value );
         } catch( Exception e ) {
             throw new JGivenWrongUsageException(
-                    "Error while trying to generate the description for annotation " + annotation + " using DescriptionGenerator class "
-                            + tagConfiguration.getDescriptionGenerator() + ": " + e.getMessage(), e );
+                "Error while trying to generate the description for annotation " + annotation + " using DescriptionGenerator class "
+                        + tagConfiguration.getDescriptionGenerator() + ": " + e.getMessage(), e );
         }
     }