+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# v0.9.2
+
+## New Features
+
+* Cases can now have custom descriptions by using the new annotation `@CaseDescription` [#139](https://github.com/TNG/JGiven/issues/139).
+
+# v0.9.1
+
+## New Features
+
+* Global formatter now apply also to subtypes of the types they are configured for [#137](https://github.com/TNG/JGiven/issues/137)
+
+## Fixed Issues
+
+* Fixed an issue with TestNG that when using a data provider and one case fails the TestNG status was still reported as success. [#138](https://github.com/TNG/JGiven/issues/138)
+
 # v0.9.0
 
 ## Backwards incompatible changes regarding placeholder variables
@@ -39,6 +55,12 @@ given().a_step_with_a_$placeholderVariable_with_spaces( "foobar" );
 ## Fixed Issues
 
 * Special characters like parentheses are not treated as part of a placeholder name anymore [#118](https://github.com/TNG/JGiven/issues/118)
+
+# v0.8.4
+
+## Fixed Issues
+
+* Fixed in issue with TestNG introduced with v0.8.3 that a failing case does not lead to failed TestNG status [#138](https://github.com/TNG/JGiven/issues/138)
 
 # v0.8.3